% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cancer}
\alias{cancer}
\title{Cancer data}
\format{A data frame with 52 observations and 9 variables:
\describe{
\item{ptno }{Patient number}
\item{treatm}{Treatment}
\item{sex}{sex}
\item{height}{height of patient in cm}
\item{weight}{weight of patient in kg}
\item{comorb}{Does the patient have one or more comorbidities}
\item{painbl}{Pain at baseline on a 0-100 VAS scale}
\item{painfu}{Pain at follow up on a 0-100 VAS scale}
\item{race}{ethnicity of the patient Levels: Asian Black Caucasian}
}}
\usage{
cancer
}
\description{
Cancer data
}
\examples{
data(pain)
}
\keyword{datasets}
